[refactor] rename base files (#30965)
## Summary & Motivation

This is more just correcting a "mistake" when originally breaking up the larger files. For symmetry, I called the file containing the core class of each of the categories (PartitionsDefinition, PartitionMapping, PartitionsSubset) "base.py", but I realized that this is kinda annoying when grepping for the file that contians the definition and is in general less clear.

Renames these files to the name of the class they define.

Also renames `time_window_partition_mapping.py` to `time_window.py` to align with the naming conventions elsewhere in this folder.

## How I Tested These Changes

## Changelog

NOCHANGELOG