Disable agent heartbeat check on serverless ci deploy calls
Summary:
This fixes the remaining blocker to moving serverless over to use the new github action templates by default, which was the initial deploy failing while it was waiting for the agent to initially heartbeat.

The old github action worked around this by detecting the first github action run and bumping up the timeout for that case, which is fairly hacky.

This check was intended to quickly error out for users who forgot to set up a hybrid agent before running CI/CD or misconfigured their agent, which should not be an issue for serverless deployments since the agent is automatically spun up for you.