Split up loading variable in WorkspaceContext (#31016)
## Summary & Motivation

After adding assets to the WorkspaceContext we changed the loading
variable to include the loading of assets. This is making non-asset
related pages wait for assets and in the case of kraftheinz (over bad
internet) their page load time on a fresh load (no indexeddb cache) is
like 5 minutes. I added `loadingNonAssets` which is equivalent to the
previous loading state and `loadingAssets` so that asset pages only
depend on assets loading.

## How I Tested These Changes

jest. local app proxy