[ui] Ensure asset check evaluations are sorted by timestamp (#31008)
## Summary

Right now, asset check executions are displayed in the order they are
returned from the db, which uses ids for pagination/sorting. In rare
cases it appears that ids may not match with evaluation timestamps. This
PR ensures that evaluations rendered on the page will always show the
newest evaluation first.