[ui] Update the asset event filters to fix type filtering
Context:
https://dagsterlabs.slack.com/archives/C03CCE471E0/p1751466532317749

This is a bit odd - we present both “Type” (Materialization / Observation) and “Status” (Success / Failure) filters, but under the hood these map to three event types in our system, and Status only exists in the new Observe UI.

This fixes support for the Type filter, which was being ignored, and allows you to use the two filters together. Type=Materialization gives you `FAILED_TO_MATERIALIZE` and `MATERIALIZATION` events, and adding Status=Success filters that to just the latter. There is no “failed observation” event, and filtering to that yields no results.

FE-884