[refactor] Simplify AssetLayer (#30990)
## Summary & Motivation

Wow was this harder than expected. There were so many things in this class that did absolutely nothing, and so many things that looked like they did nothing, but actually encoded subtle behavioral things. Wowee.

Anyway, this class had a bunch of complexity for no good reason. Any time you see a big bundle of dictionaries, you know something has gone horribly wrong.

I've restructured this class completely and pared down its methods to be as simple as possible. All of the complexity is shoved into the implementation of the `_asset_keys_by_inner_input_handle` property, which I've simplified as much as possible, and added comments as necessary. Perhaps I've been looking at it too long but I think it's actually fairly understandable at the moment!

The main bit of complexity that this class handles (excuse the pun) is that there are technically multiple NodeOutputHandles / NodeInputHandles that correspond to an edge in the graph if you are using graph-backed assets. There are "outer node handles" which correspond to the graph output, and "inner node handles" that correpsond to the underlying op. These can be nested arbitrarily deeply in principle.

The old regime did not do a great job of distinguishing between these different types of node handles, which I think I have overall cleaned up a bunch.

In terms of net LoC, it's only 20 lines fewer, but:
- There are a lot more comments now!
- This is substantially simpler structurally, so someone can come in and work on this without going through all the pain that I did for this PR
- Once we clean up AssetsDefinition usage in a similar way, this will all get even slimmer.

## How I Tested These Changes

## Changelog

NOCHANGELOG