[ui] Update the asset event filters to fix type filtering (#31030)
Context:
https://dagsterlabs.slack.com/archives/C03CCE471E0/p1751466532317749

This is a bit odd - we present both “Type” (Materialization /
Observation) and “Status” (Success / Failure) filters, but under the
hood these map to three event types in our system, and Status only
exists in the new Observe UI.

This fixes support for the Type filter, which was being ignored, and
allows you to use the two filters together. Type=Materialization gives
you `FAILED_TO_MATERIALIZE` and `MATERIALIZATION` events, and adding
Status=Success filters that to just the latter. There is no “failed
observation” event, and filtering to that yields no results.

I also fixed the icons used in the submenus of this filter picker since
they were all "job"

I also hid the "Status" filter for non-materializable assets, since
there is no Failure event for observations

FE-884

## How I Tested These Changes

I tested this by visiting an asset and making selections in the filter
menu and confirming that I see the right params being sent to the
graphql resolver for different selections.

## Changelog

[ui] The "type" filter on the asset events view now behaves as expected
in all scenarios.

Co-authored-by: bengotow <bgotow@elementl.com>