[docs snippet tests] Clear and regenerate snippet folders (#31041)
## Summary

The docs snippets tests can be frustrating, not least because it's
unclear which snippets are actually up-to-date and which are artifacts
of past versions of the tests which were never cleaned up.

This PR removes and regenerates the docs snippet folders completely when
updating snapshots for dg/components tutorials. This prevents old
snippets from sticking around.

Moves a few tests which have both generated and non-generated snippets
to a `/generated` subfolder.