[Asset graph] Workaround fix for long loading times (#31048)
## Summary & Motivation

Currently the asset graph is sometimes ending up in an infinite loading
state.

From my repro I was able to see that we were marking asset layout
requests as cancelled (which happens in the return of layout effect) yet
another layout effect does not fire... this means we end up throwing
away the layout without scheduling a new layout. While debugging I saw
we were throwing it away even though the layout cache key hasn't changed
so add a check of the layout cache key as a further check that the
layout should really be cancelled.


## How I Tested These Changes

local testing