[refactor] Update all imports in test files (#31059)
## Summary & Motivation

Updates all imports in our test code to use `import dagster as dg` instead of `from dagster import xyz`.

This removes a bunch of unnecessary LoC and also makes it less annoying to bring in new APIs when adding tests to existing files, something I've been bit by a lot. In general the current flow is:

1. start typing a new import (say, AssetCheckSpec)
2. you notice it's not imported
3. you let your editor auto-import the symbol
4. because dagster is a "first party library", it uses the hyper-specific import path
5. you end up with a bunch of lines of code that look like:

![Screenshot 2025-07-03 at 3.54.34 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/cTPQ7oSxo9qxXgkzpnaK/13ce8016-1b0b-4ec5-bb10-22cce1a911b3.png)

Now, the flow is:

1. type `dg.AssetCheckSpec`

This was done with a janky script that parses the AST etc., and was brought to you by the wisdom of AI and definitely misses some stuff. If you look at some of the updated files, you'll see that some of them still have a few residual "from dagster import ..." imports within them. That's certainly not ideal but figured that this is enough of a step forward (-2.5k LoC) that we can live with a few remaining stragglers and update those as we see fit.

## How I Tested These Changes

Everything is broken at the moment.

## Changelog

NOCHANGELOG