[perf] Back global search by workspace context (#31065)
## Summary & Motivation

Back global search by the workspace context. The SearchPrimaryQuery
which was largely duplicative of the WorkspaceContextQuery except that
it wasn't split by code location and it was never refreshed so it could
become stale easily making it just overall worse for powering search.
This should significantly reduce memory usage of the app for users with
large non asset based workspaces.

## How I Tested These Changes

App proxy, existing jest tests.