Make dagster-dg-cli depend directly on dagster
Summary:
Right now dagster-dg-cli only brings in dagster if you use the local extra. This was I think intended to preserve option value for a more lightweight version of dg. But in practice it just makes it easy to accidentally install dg with invalid dependencies. For example if you do pip install -U dagster-dg-cli it won't realize that it also needs to upgrade dagster and will leave things in a broken state with confusing error messages.

This also means we don't need validate_dagster_availability anymore.