[ADOPT-1614] Better error message when scaffolding shim components that already exist (#31071)
## Summary & Motivation

The check we were using was overly-specific. While it's true that most scaffolding operations result in a directory being created, for things like assets and resources, a single file is created instead. So we weren't catching this error early enough, resulting in an ugly stack trace output if you tried to scaffold an (e.g.) asset multiple times with the same path.

## How I Tested These Changes

unit test

## Changelog

NOCHANGELOG