Use a generic return type in AutomationCondition.replace() (#31106)
## Summary & Motivation

This introduces a generic TypeVar to represent the "new" Automation
condition in `AutomationCondition.replace()`, and uses that TypeVar in
the return type as well to give the caller more information about what
class of `AutomationCondition` might be returned.

This solves the problem where:

```python
dg.AutomationCondition.eager().replace(
    "newly_updated", dg.AutomationCondition.data_version_changed()
).with_label("eager_respecting_data_version")
```

fails typechecking. This happens today because `replace` always returns
`AutomationCondition`, and hence we don't know that `with_label` will
exist on it (that only exists on `BuiltinAutomationCondition`).

But we should be able to know that `with_label` exists; `replace` always
returns either `Self` or the type of `new`, both of which are indeed
`BuiltinAutomationCondition`.

## How I Tested These Changes

Added a unit test. The test fails `make pyright` before this change, and
passes it after this change.

## Changelog

Updates AutomationCondition.replace type signature to provide callers
more information about the returned AutomationCondition.