Do not initialize repo unnecessarily in _get_code_pointer_dict_from_python_pointer_opts impl (#31173)
## Summary

The impl of `_get_code_pointer_dict_from_python_pointer_opts` needs to
determine the repository name in order to produce its return dict - to
do so it needs to fully initialize a repository.

This results in several very silly cases where we initialize a
repository just to reinitialize it a few lines later!

https://github.com/dagster-io/dagster/blob/master/python_modules/dagster/dagster/_cli/job.py#L464-L469

Now that most entrypoints use `Definitions` or `LazyDefinitions`, repo
names are typically just the singleton, so short-circuit in these cases.