Create an external asset for upstream keys referenced by asset checks that are not in the graph
Summary:
The comment here claims that we covered the case in the new unit test, but we did not. Now we do.

Test Plan:
New test case

Changelog:
Fixed an issue where asset checks referencing asset keys that did not exist in the asset graph did not appear in the Dagster UI.