[partition-context] partition_loading_context guard (#30966)
## Summary & Motivation

Now we (finally) get to the meat of this stack. We have hit a glut of performance and runtime error landmine type issues due to the somewhat silly way all the partition-related methods work.

Basically all of our public methods (and private ones), take two optional parameters: `current_time` and `dynamic_partitions_store`. The curent time argument is only relevant for time-window partitions, and even then most methods function fine without it being set. However, if it is not set, then it becomes very easy to hit perf bottlenecks, specifically around getting the last partition key available. This of course depends on the current time, and so if we're constantly invalidating the cache then this means we have to do a moderately-expensive calculation many times. There are also correctness issues that are in principle possible here, where the current time passes a partition boundary during the course of evaluation. The performance thing in general has bitten us a lot, because it's impossible for us to enforce that the current time is threaded through everywhere, as again these parameters are optional. There are also cases where properties or things like `__len__` are actually dependent on the current time but have no way to define arguments.

Similarly, the dynamic partitions store argument is only used for dynamic partitions definitions, and we've had multiple cases where nobody had attempted to use certain code paths with a dynamic partitions defintiion for quite awhile, leaving a landmine where once someone finally did, they hit an error because the argument was not properly threaded through.

This PR creates a context var that stores a PartitionLoadingContext that is used across all the partition-related functions. Whenever a method is invoked that has these arguments, it *adds* the passed-in context to whatever context was already set externally.

This means that if you're defining an operation that is going to be doing a lot of partition-related operations (e.g. a backfill or declarative automation evaluation), as long as you set this context guard outside of your main computation, ALL partition-related operations will be guaranteed to have the current_time / dynamic_partitions_store set. This effectively eliminates an entire class of bugs and performance issues -- as long as you set this up once, it's impossible for inner invocations to mess this up, no matter how obscure the invocations are.

## How I Tested These Changes

## Changelog

NOCHANGELOG