[da] Fix bug with serialization of user-defined conditions (#31102)
## Summary & Motivation

Fixes an issue when attempting to serialize snapshot data as shown in the provided test case.

The core issue is that the `is_serializable` property was not properly recursing over the children of the referenced dep conditions.

## How I Tested These Changes

unit

## Changelog

Fixed a bug that could cause code locations to fail to load when a custom python AutomationCondition was used as the operand of `AutomationCondition.any_deps_match()` or `AutomationCondition.all_deps_match()`