[definitions] rm assertion for unresolved definitions, add merge_unbound_defs (#31178)
## Summary

Allows `Definitions.merge` to merge objects which have previously had
the asset graph built, which currently triggers a check error.

The check error is currently blocking cross-component access in #30446.
For cross-component accesses, a component triggers the a `Definitions`
object to be built for another component, and subsequently accesses
properties of that `Definitions` object. If that property requires
resolution, such as the list of asset keys, it will cause the subsequent
merging of `Definitions` objects at the folder level to fail this check.

Adds a new internal `merge_unbound_defs` which contains the assertion.

## Test Plan

New unit tests.