[partition-mapping] Fix issue causing us to not create an AllPartitionsSubset (#31218)
## Summary & Motivation

After the big refactor, this condition was always going to be False because we never manually pass in these values (they're available via the context).

This updates the conditional to inspect the context for the presence of these things.

## How I Tested These Changes

Added a test that failed before this change and passes now

## Changelog

NOCHANGELOG