Raise an exception if you try to retry a run with no steps or assets or checks
Summary:
It's possible for a run to do all of its intended work but still fail (for example, it finishes all its steps and then fails during run cleanup, or completes all its asset materializations but then the step fails). In this case, the retry logic will return an empty list of steps/assets/checks, and then either fail with an inscrutable error related to io managers, or worse, sometimes try to run every step (!) This fixes that by detecting the error and raising an informative exception.

> Insert changelog entry or delete this section.