[observe] Add a utility to format durations (#31266)
## Summary & Motivation

As titled adds a utility that given a duration returns either a single
unit or 2 units (defaults to 1) to represent the duration.

I'm not sure if these are the right "rules" but I made it easy to modify
the behavior via the `maxValueBeforeNextUnit` prop which is a dictionary
mapping each unit to the max number we should show for it before instead
using the next unit.

## How I Tested These Changes
jest