[dg cli] docs for component by type (#31251)
Add a cli entry point for getting the details of a specfiic component
type

## How I Tested These Changes

added test

example non-json output 
```
$ dg docs component dagster.UvRunComponent

Type: dagster.UvRunComponent
Description: Represents a Python script, alongside the set of assets or asset checks that it is responsible for executing.

Accepts a path to a Python script which will be executed in a dagster-pipes subprocess using the `uv run` command.

Example:
```yaml
type: dagster.UvRunComponent
attributes:
  execution:
    path: update_table.py
  assets:
    - key: my_table
```
```