[BUILDKITE_FOLDER_BRANCH=buildkite][NO_SKIP] Add buildkite_shared folder to copied folders (#31205)
https://github.com/dagster-io/internal/pull/16979 moves
CommandStepBuilder to a shared directory in .buildkite and this PR
ensures that directory is copied over. Also, with that move to a shared
directory, I ran into some issues with different ruff expectations in
the internal ruff configuration vs dagster configuration on the
.buildkite directory. I chose to exclude .buildkite from being
considered by the dagster ruff step.

buildkite did fail on this build, but I reviewed the failures and have
noted that the few steps/tests are unrelated