[backfill] Fix issue that could cause incorrect backfill requests (#31285)
## Summary & Motivation

This is pretty hard to reproduce in unit tests, but we generally need to be pretty strict about the cases where we launch a child with its parent because our run building machinery is strict about the way in which it builds runs.

The core thing here is that in some cases, we have a target parent subset of: `[a, b]`, and that maps to a target child subset of `[a, b]`. The previous logic would see this, see they had the same partitions definition, all that good stuff, and decide that it was safe for them to get materialized in the same run. However, that target parent subset that we got from the bfs iteration would not actually be all of the target parent partitions that we planned to materialize, and we would end up requesting:

- parent: [a, b, c, d, e]
- child: [a, b]

Then, our run machinery would not be able to group parent and child into the same run because they target different sets of partitions.

## How I Tested These Changes

local

## Changelog

Fixed an issue that could cause asset backfills to request downstream partitions at the same time as their parent partitions in rare cases.