[dagster-tableau] Yield AssetObservation events for sheets and dashboards (#31260)
## Summary & Motivation

This PR changes the behavior of `refresh_and_poll` so that we emit asset
observation events for sheets and dashboards.

In Tableau, only data sources with extracts can be materialized
(refreshed). Other data sources are live connections, and
sheets/dashboards are views of their underlying data sources.

Because of that, sheets and dashboards are not directly materialized via
Dagster. We can observe these assets and update the metadata accordingly
as a workaround.

## How I Tested These Changes

Updated tests with BK.

## Changelog

[dagster-tableau] Tableau sheets and dashboards are now being observed
and instead of being materialized when using `refresh_and_poll` inside
the `@tableau_assets` asset decorator.