[dagster-tableau] Support dashboards and data sources with hidden sheets (#31268)
## Summary & Motivation

When fetching workbook details via the Tableau Metadata API (GraphQL),
sheets for which the LUID is missing are considered hidden. Previously,
we were discarding them, but also discarding their data sources in the
process.

In Tableau, dashboards can contain hidden sheets - we want to make sure
we represent these dashboards as it should, with their upstream data
sources, even if sheets between these dashboards and upstream data
sources are hidden.

This PR updates the code so that we represent upstream data sources as
direct deps of these dashboards in the asset lineage.

## How I Tested These Changes

Updated tests with BK

## Changelog

[dagster-tableau] Dashboards containing hidden sheets are now correctly
linked to upstream data sources.