Mock SHELL in install completion tests (#31295)
When we refactored our divergent buildkite step builders into a single
consolidated step builder, this test regressed.

The refactor cut down on implicitly passed env vars and I think that
caused this particular test to start failing. I've had trouble running
down the exact env var because, well, it wasn't explicitly being set...

But given what this test is doing, I think it's better that we mock our
shell and actually test the different conditions here. Instead of having
a test whose assertions are tied to the environment in which it's
running.