go through parent keys in a consistent order in the asset backfill traversal
Summary:
This makes dry runs more consistent/reproducible. Ideally the bfs traversal would also be sorted in a similar fashion but that is probably a larger change.

test that is still having trouble with grouping

Summary:
I'm still seeing the following graph have some trouble with grouping - a child has two parents with matching partitions definitions, but we are only materializaing a smaller subset of one of the parents. Neither of the parent mappings produces a failure reason, but we are materializing a different subset of one of the parents, so parent A and child get grouped into one run. Still digging into what exactly is going wrong.

> Insert changelog entry or delete this section.