[backfill] targeted fix for can run with parent logic (#31315)
## Summary & Motivation

We have some very confusing behavior at the moment, where in cases where an asset has multiple parents, it's possible for the system to determine that the entirety of a downstream asset may be able to run with one of its parents, and then when another parent is factored in and we determine that we cannot run with part of that parent, we don't accurately remove all assets from the "passed subset".

The core idea here is that we believed we could execute the given subset with its parent because all of the subset would be included in the resulting run. As soon as anything influences this fact, the core assumption breaks down and the partitions get sent to separate runs.

This is a super-targeted fix for this behavior and realistically this entire codepath should be reworked.

## How I Tested These Changes

## Changelog

NOCHANGELOG