Initial docs skeleton (#31331)
## Summary & Motivation

I decided to put this at the bottom of the stack to organize all the capabilities we are building so it is clear what is going on. That way it is not just a bunch of random script but a well-documented CLI.

A few features:

* The ability to parse and lint docstrings at the symbol, package, or global level. This will be very useful for writing docstrings and is a couple orders of magnitude faster than the current docs buidl toolchain
* Provides the tools to allow us to keep @public-annotated exported and .rst entries in sync. The global here is that every @public class and function is in rst files in docs and every class or function in .rst has @public. That way any developer in the codebase knows when they are dealing with a public artifact, and we ensure that every public artifact has docs, and this can be enforced with tools.

## How I Tested These Changes

BK as well as manual testing.