pass through asset selection when launching a sensor/schedule preview (#30987)
## Summary & Motivation
If a sensor specifies an asset selection, or if the RunRequest returned
by a sensor or schedule evaluation specifies an asset selection, we
should pass that asset selection through to the run execution in the dry
run sensor/schedule ui flow.

In fixing this, I noticed that we are not passing through asset check
selections either, I'll fix that in a stacked PR (#30998). I also think
the UI could use some help to display the selected assets better, so
i'll take a stab at that too

fixes #27647 


## How I Tested These Changes
tested in local dev. I looked at writing a unit test, but it seemed like
the mocks for the sensor dry run page hard code the execution parameters
when launching a run, so i dont think i would be able to test that the
results from evaluating the sensor are correctly transformed into the
execution parameters. Maybe i'm missing another test file where testing
this would be more appropriate.