[dagster-databricks] Enrich cluster config vars (#31094)
## Summary & Motivation

The list of configuration vars for new cluster provisioning in the
dagster-databricks library was lacking with respect to what is offered
in newer versions of the Clusters API or the Databricks SDK. More
specifically, the following keys were added to the cluster
configuration: `azure_attributes` and `gcp_attributes` (to go alongside
existing `aws_attributes`), `data_security_mode`, `single_user_name`,
and additional init script configuration for Azure and GCP.

This is not an exhaustive list of what is available in the API, but
brings parity for GCP and Azure infrastructure compared to what was
available for AWS in prior versions, and adds crucial variables related
to Unity Catalog operation.

Additionally, some of the config enums were mapped directly to their
equivalent enums in the SDK, as opposed to being a copy/pasted list of
values, which could potentially change upon newer versions of the SDK.

## How I Tested These Changes

Tested locally with existing tests, and with creation of a new cluster
via the databricks step launcher on Azure infrastructure.