[ui] Fix subpixel rendering in Automations list (#31473)
## Summary & Motivation

@salazarm noticed that the list items in the new Automations list were just rendering at the estimated size, instead of being measured at the intended size. This is because of the `height: 100%` hack that I put in place to deal with subpixel rendering causing row box-shadows to be cut off by tanstack virtualizer's rounding off subpixel measured values. With `height: 100%`, the bug appeared to be fixed, but it was just because the row occupied the estimated size, the measurement matched that, and nothing changed.

We identified the small "caption" text as being the source of the subpixel height, due to Blueprint's body default line-height being a strange magic number, and our not currently having any specific line-height set on certain text components.

This PR adds line-height values to these components (for which we will need to keep an eye out for visual bugs) and just changes the text component usage in the Automations list to use `BodySmall` and `MonoSmall`, which comments indicate are non-deprecated text components.

The result is that the list items no longer have subpixel rendering, the item borders render properly, and the measurement is correct.

I also modified the row estimate to be 62px instead of 64px, which is more accurate and leads to less render jumping during scroll.

## How I Tested These Changes

View Automations list, verify that list items render correctly, including box-shadow borders and during scroll.