[ui] Undo ListItem right element top/bottom spacing (#31483)
## Summary & Motivation

The top/bottom spacing I just added to the right elements in ListItem are excessive when applied in addition to the `align-self` behavior.

I think probably if top/bottom padding is desired on `right` items, it should be added on the `right` element itself, rather than the `ListItem`. But hopefully we just generally shouldn't need it.

Before:

<img width="828" height="72" alt="Screenshot 2025-07-31 at 12 04 27" src="https://github.com/user-attachments/assets/32f4de54-17c3-466e-8555-9e1c104f8bb9" />


After:

<img width="827" height="56" alt="Screenshot 2025-07-31 at 12 04 34" src="https://github.com/user-attachments/assets/44a5078e-8178-453a-bf69-5b29c83b2449" />


## How I Tested These Changes

View Home, verify that failed jobs list looks correct.