[ui] Additional export to support billing charts (#31472)
## Summary & Motivation

This is a tiny PR exposing a utility we use for LineCharts so that I can
reuse it within the billing charts. On the billing page we have a Line
chart, but it has a custom timeframe (the billing period), different
tooltips, etc. so this is the only part that it reuses.


https://linear.app/dagster-labs/issue/OPER-2230/billing-page-design-updates-for-self-managed-upsells

Co-authored-by: bengotow <bgotow@elementl.com>