[RFC] Set default label for AutomationCondition to name
Today, AutomationConditions have names and labels; names are always set and are determined by the AutomationCondition class itself; labels are customizable and are optional.

However, this leads to a few footguns. `replace()` targets AutomationConditions by label, but that means that some conditions are unexpectedly not replaceable (see https://github.com/dagster-io/dagster/issues/25439).

So here's a proposed fix: change AutomationCondition's label to be non-optional, and have it default to the name. Then if you set a label, that overrides it instead.

All existing tests pass, as does typechecking, and a new test case is added demonstrating the functionality.