[dagster-dbt] dbtf compat (#31337)
## Summary & Motivation

This is a set of changes that allow the dagster-dbt library to work for users using the new dbt Fusion engine.

There are two main parts to these changes:

**Allow dagster-dbt to function without the `dbt-core` python package**

the dbt Fusion engine is written in rust, and so does not have a corresponding python installation. the previous implementation of this package required that dbt-core was installed. now we:

1. remove the dbt-core requirement from dagster-dbt
2. keep track of if we have dbt-core installed
3. if we do, use the available methods for things like resolving dbt selection strings
4. if not, we invoke dbt cli commands 

we also hard-code certain types if dbt-core is not installed (e.g. NodeStatus, TestStatus).

**Handle different log output formats**

the log messages that are produced from dbt Fusion commands are broadly in a similar format, but have some minor differences that need to be accounted for. In particular, all tests / model executions are represented with a "NodeFinished" event type, whereas the dbt-core log messages had a bit more variety.

to handle this, we create separate subclasses for parsing these messages depending on what cli is producing them.

## How I Tested These Changes

## Changelog

> Insert changelog entry or delete this section.